requirements.txt
darkdetect==0.7.1
firebase_admin==6.5.0
PyQt6==6.7.1
PyQt6_sip==13.8.0
pywin32==306
QtAwesome==1.3.1
reportlab==4.2.2
svglib==1.5.1
PyQtDarkTheme-fork==2.3.2
PyQt6-WebEngine==6.7.0
GitPython==3.1.43